// packages/web/src/shared/api/generated/model/createTransferTransactionDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Finnn API
 * Backend API for Finnn finance tracking
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTransferTransactionDto {
  fromAccountId: string;
  toAccountId: string;
  amount: string;
  toAmount: string;
  description?: string;
  date: string;
}